controller/requirements.txt
# Deis controller requirements
#
Django==1.6.11
django-cors-headers==1.0.0
# required for south migrations
django-fsm==2.2.0
django-guardian==1.2.5
django-json-field==0.5.7
django-auth-ldap==1.2.5
djangorestframework==3.0.5
docker-py==1.5.0
gunicorn==19.3.0
marathon==0.6.15
paramiko==1.15.2
psycopg2==2.6.1
python-etcd==0.3.2
python-ldap==2.4.19
PyYAML==3.11
semantic_version==2.4.2
simpleflock==0.0.2
South==1.0.2
static==1.1.1
astakosclient==0.16
django-longerusernameandemail==0.5.7
